requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob==12.14.1
patool==1.12
pyunpack==0.3
# no idea
azure-core==1.26.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
EasyProcess==1.1
entrypoint2==1.1
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
pycparser==2.21
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12